+++ src/users/serializers/user_serializers.py
@@ -11,10 +11,3 @@ class UserSerializer(serializers.ModelSerializer):
         model = User
         fields = ["id", "email", "name", "gender", "phone_number", "profile_image", "prefer_service", "prefer_location"]
         read_only_fields = ("id", "email")
-
-
-class UserInfoSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ["id", "email", "name", "profile_image", "is_expert"]
-        read_only_fields = ("id", "email")
